Add git-lfs to host dependencies
The problem I'm seeing arises from the way conda-build checks out
a bare copy of tracktable-data, then tries to clone that.  It is
failing to notice that there are Git LFS files in the repository
that it needs to fetch.

There's a comment in https://github.com/conda/conda-build/issues/4882
suggesting that we add git-lfs to either the build or host section
of the build recipe.  I've already got it in the build section and
that doesn't appear to be enough.  This commit adds it to the host
section too.