Convert to new conda-forge recipe format
Use rattler-build for builds instead of conda-build.

This commit also changes the recipe to pull from GitHub release
archives instead of the repository itself to save on conda-forge
bandwidth.

Also also, we pull in the tracktable-docs release so that the
notebook-converting script can do its thing.  This issue will go
away once we handle the re-merge between tracktable and tracktable-docs.