Disable libcpp availability check in build.sh
There is a new mechanism in libc++ that checks to see when features
became available (in the underlying OS distribution) compared to
which version of clang is running.  This was firing on something
(from_chars, I think, related to floats) in boost::uuid.

The conda-forge knowledge repository has this to say:

"""
However, since conda-forge ships its own (modern) libcxx we can ignore
these checks because these symbols are in fact available. To do so, add
_LIBCPP_DISABLE_AVAILABILITY to the defines. For example

CXXFLAGS="${CXXFLAGS} -D_LIBCPP_DISABLE_AVAILABILITY
"""

I have added that flag to the C++ compiler arguments in build.sh.